clang-tidy kapısını gerçekten çalıştır ve ağacı temizle
`make check` clang-tidy'yi atlıyordu. Makefile `command -v clang-tidy` diye
soruyor, Homebrew'un LLVM'i keg-only olduğu için "kurulu değil" cevabını alıyor
ve SKIPPED yazıp geçiyordu — araç iki dizin ötede duruyorken. Soru artık
`scripts/tidy.sh --var`'a soruluyor; aracı arayan kod zaten oradaydı. Atlayan
bir kapı hiç çalışmamış kapıdır (CLAUDE.md 6.2).

Kapı açılınca 3719 bulgu çıktı. Hepsi kapatıldı. Gerçek olanlar:

- ÜÇ YERDE İŞARETÇİ SIRALANIYORDU — `command/registry.cpp` katalog parmak izini,
  `core/settings.cpp` proje ayarlarının katlamasını, `processing/registry.cpp`
  araç ağacını böyle kuruyordu. Karşılaştırıcılar anahtara bakıyordu ama adres
  kabını yerinde sıralamak, bu programın iki makinenin bit bit anlaşacağı
  iddiasının (Article 2.5, §7.3) yanında durulacak bir şekil değil. Üçü de
  indeks sıralamaya çevrildi.
- `app/main.cpp`: `check()` rapor ediyor, dönmüyor. Bir sonraki satır az önce
  şikâyet ettiği null'ı dereference ediyordu.
- `app/layout_designer.cpp` kullanıcıya yanlış yer tutucuyu söylüyordu.
- `Banner::addAction` `QWidget::addAction`'ı gizliyordu → `addButton`.
  `GridDelegate::setTheme` sanal olmayan tabanı gizliyordu → taban `virtual`.
- `io/dxf_reader.cpp`: `PendingInsert` her vektör büyümesinde koca bir DXF
  varlığını kopyalıyordu; artık işaretçinin arkasında ve taşınması atmıyor.
- `kentos_docgen`'in `main`'i istisna sızdırıyordu. Yarım yazılmış bir referans,
  `ci-gate-docs.sh`'in karşılaştıracağı şeydir.
- `core::LayoutItem` nesne başına 39 bayt dolgu harcıyordu (en iyisi 7).
- İki kurucuda sanal çağrı, iki denetlenmemiş optional, üç taşma riski,
  bir yanlış yuvarlama, iki null dereference.

`EntityFlag` artık `: unsigned`. Bayraklar ondan katlanan her maskeyi işaretli
bir ifade yapıyordu; BAYT hâlâ `EntityTable::flags` ve dosya biçimi onu `u8[]`
olarak yazıyor — değişen yalnız numaralandırıcıların kendi türü (model.md R6, R7).

`scripts/tidy.sh` bulguyu konumuna göre süzüyor: `/src` dışındaki bir bulgu
üzerinde çalışabileceğimiz bir bulgu değil. Betiğin kendi yorumu bunu zaten
söylüyordu; clang-tidy'nin `HeaderFilterRegex`'i çözümleyici bulgularına
ulaşmadığı için — çözümleyici her satıriçi işlevi kendi giriş noktası sayıyor,
libpqxx'in `result_iter`'ı canlı örnek — politika koda taşındı. Her koşut çocuk
kendi çıktısını süzüyor, çünkü tek akışa yazan birkaç süreçte bir tanı ile
notlarının birlikte gelmesi garanti değil.

`make check` çıkış 0, ctest 45/45, kentos_tests 754/754.

+++ src/io/src/dxf_writer.cpp
@@ -475,8 +475,9 @@ private:
         common(sp, e);
         sp.normalVec = DRW_Coord(0.0, 0.0, 1.0);
         sp.degree    = d.degree;
-        sp.flags     = (d.closed ? 1 : 0) | (d.periodic ? 2 : 0) | (d.rational ? 4 : 0) |
-                   (d.planar ? 8 : 0) | (d.linear ? 16 : 0);
+        sp.flags =
+            static_cast<int>((d.closed ? 1U : 0U) | (d.periodic ? 2U : 0U) |
+                             (d.rational ? 4U : 0U) | (d.planar ? 8U : 0U) | (d.linear ? 16U : 0U));
         const auto cx = geo.ring_xs(span.first);
         const auto cy = geo.ring_ys(span.first);
         for (std::size_t i = 0; i < cx.size(); ++i)
@@ -586,7 +587,9 @@ private:
         }
         const auto fill = [&](DRW_Dimension& out, int type_bits) {
             common(out, e);
-            out.type = type_bits | (d.user_text_position ? 128 : 0) | (d.ordinate_x ? 64 : 0);
+            out.type =
+                static_cast<int>(static_cast<unsigned>(type_bits) |
+                                 (d.user_text_position ? 128U : 0U) | (d.ordinate_x ? 64U : 0U));
             out.setTextPoint(text_at);
             out.setStyle(d.style);
             out.setText(d.override_text);
